add single-export files
